Change message tag logic
So far, `None` meant any tag is acceptable. From now on, tags must match
exactly. Per default, the tag of and message is `"default"`. This should
make it possible for a subscriber to push data through the network and
receive the results via the callback mechanism without other subscribers
listening in, via setting a private tag. This will require modifications
to the `Exporter` and the `forward()` wrapper.